+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add publish\_to: none when localizing
 - Add .gitattributes file
 - Add --git-ref option to localize-refs for custom git refs
+- Add test for updating .gitignore with missing .gg entries
 
 ### Changed